requirements.txt
arandr==0.1.11
astroid==2.14.2
asttokens==2.2.1
automationhat==0.2.0
beautifulsoup4==4.11.2
blinker==1.5
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.2.16
distro==1.8.0
docutils==0.19
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==2.2.2
fourletterphat==0.1.0
gpiozero==1.6.2
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==0.0.0
logilab-common==1.9.8
lxml==4.6.3
MarkupSafe==2.1.1
mccabe==0.7.0
microdotphat==0.2.1
more-itertools==8.10.0
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
numpy==1.23.2
oauthlib==3.2.2
olefile==0.46
pantilthat==0.0.7
parso==0.8.3
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.16.2
pycups==2.0.1
pygame==1.9.6
Pygments==2.14.0
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.7.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.2.1
python-prctl==1.7
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
roman==3.3
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.8.1b0
sense-hat==2.4.0
simplejpeg==1.6.5
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
thonny==4.0.1
toml==0.10.2
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing_extensions==4.4.0
unicornhathd==0.0.4
urllib3==1.26.12
v4l2-python3==0.3.2
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.8.1
